Added documentation. Minor refactor to separate concerns into separate python files.
